Adding PR from parent repo:
I implemented my Kubernetes cluster using this repository. This is an amazing and helpful guide. However, when I installed the Flannel CNI plugin for my cluster, I encountered the error: Error registering network: failed to acquire lease: node "node-1" pod cidr not assigned.

To resolve this, I added the --allocate-node-cidrs=true option to the kube-controller-manager.service to automatically allocate subnet IP ranges for the nodes. This is helpful for installing CNI plugins later.

https://github.com/kelseyhightower/kubernetes-the-hard-way/pull/815/commits/f5069ce07d7b2624194c23aa933e84e59a05219d